feat(compress): accept valid zstd windows
## What

- Allow zstd decompression to accept valid frames whose window is larger than the requested decoded output limit while still enforcing the caller-provided decoded size.
- Move codec-specific compression tests out of the root registry tests and into the owning codec packages.
- Keep the zstd window regression in the zstd package test, using the upstream zstd encoder only as test setup.

## Why

- Some valid non-single-segment zstd frames were rejected because the decoder memory/window limit was tied to the requested decoded output size instead of the repository byte-size policy.
- Splitting codec tests keeps root compression tests focused on registration and prevents zstd-specific setup from expanding the production API.

## Testing

- `go test ./compress/...` - passed
- `make lint` - passed
- Full `make specs` was not run; validation was scoped to the changed compression packages plus repository lint.
